tests/wpt/tests/xhr/formdata/get-formelement.html

<!doctype html>
<html lang=en>
<meta charset=utf-8>
<title>FormData: get and getAll (with form element)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#dom-formdata-get" />
    <link rel="help" href="https://xhr.spec.whatwg.org/#dom-formdata-getall" />
<div id="log"></div>
<form id="form">
    <input type="hidden" name="key" value="value1">
    <input type="hidden" name="key" value="value2">
</form>
<form id="empty-form"></form>
<script>
    test(function() {
        assert_equals(new FormData(document.getElementById('form')).get('key'), "value1");
    }, 'testFormDataGetFromForm');
    test(function() {
        assert_equals(new FormData(document.getElementById('form')).get('nil'), null);
    }, 'testFormDataGetFromFormNull');
    test(function() {
        assert_equals(new FormData(document.getElementById('empty-form')).get('key'), null);
    }, 'testFormDataGetFromEmptyForm');
    test(function() {
        assert_array_equals(new FormData(document.getElementById('form')).getAll('key'), ["value1", "value2"]);
    }, 'testFormDataGetAllFromForm');
    test(function() {
        assert_array_equals(new FormData(document.getElementById('form')).getAll('nil'), []);
    }, 'testFormDataGetAllFromFormNull');
    test(function() {
        assert_array_equals(new FormData(document.getElementById('empty-form')).getAll('key'), []);
    }, 'testFormDataGetAllFromEmptyForm');
</script>
